Created lgc/include hierarchy
Part of making the LGC subdirectories more cleanly modularized.

The interface to LGC used by the front-end is now in lgc/interface, and
lgc/include is now a hierarchy of include files used from a different
LGC subdirectory, and #included with a qualified path.
For example, PipelineState.h is now in
lgc/include/lgc/state/PipelineState.h, so that it is #included as:
  #include "lgc/state/PipelineState.h"

This makes it more obvious which include files define functionality used
from different LGC subdirectories, and makes the inclusion of them work
in a way more like LLVM does. Unlike LLVM, we have two include
hierachies: we have lgc/interface/lgc for include files that make up the
LGC interface used by the LLPC front-end, and a separate
lgc/include/lgc for include files not exposed to the front-end.

A benefit of this approach is that the include paths list in
CMakeLists.txt is much simpler.

Also similarly qualified #includes of files in lgc/imported/chip.

Also re-sorted #includes now some of them are qualified.

Change-Id: I86061c066ec58d396dd71831036cf2cb69ab9f87