Put option defaults into LgcContext so driver does not need to set them
Before this commit, both the driver (compiler_solution_llpc.cpp) and
amdllpc.cpp needed to set the same set of LLVM options. With this
commit:

* The scalarizer-related options are LLPC options, so I changed the
  option defaults to what the code above currently sets.
* The other (unconditionally-set) ones now have default values set in
  LGCContext.
* I removed the amdllpc option-setting code as it is no longer needed.
  The driver code for unconditional options can be removed too.
* That will leave the options set in the driver conditionally on the
  app profile. They should be removed in the future by them becoming
  tuning options.

A further benefit of this is that the lgc utility does not need extra
command line options to behave in a way that is consistent with running
LGC as part of amdllpc.

Change-Id: I4353af365dd239cba08959b6d1e4c606d14431e3