Improved interface for building "unlinked" ELF
The front-end LLPC interface now has an "unlinked" flag in
GraphicsPipelineBuildInfo and ComputePipelineBuildInfo, so that amdllpc
can set it for a non-pipeline compile, and also set it for a pipeline
compile where the new -unlinked option is specified.

(I did not bump the LLPC version for this interface change, as it does
not break compatibility between old driver and new LLPC.)

Inside the front-end, there is a new "unlinked" flag in PipelineContext.
As well as being set by the above, it is set by
buildPipelineWithRelocatableElf.

The front-end signals to LGC that it wants an "unlinked" half-pipeline
ELF via a new flag arg in Pipeline::irLink.

That flag is stored in PipelineState, replacing the flag in LgcContext.
It is also serialized from there, which means you can write an LGC lit
test with it set.

An unlinked ELF does not have a null fragment shader added, so all of
the above subsumes the --disable-null-frag-shader option previously used
to stop a null fragment shader being added for a single shader compile
from amdllpc.

Change-Id: I6369fdaeee0fffe909a7e1ba9c514982e086791c