Defined interface for ELF linking in LGC
Up to now, ELF linking has happened in two separate bits of code in
llpcCompiler in the front-end. This commit defines (but does not
implement or use) a new LGC interface for doing it in the middle-end.
The idea is that linker code in the middle-end will use LLVM's ELF
reading/writing support, and implementation details like how the relocs
work and how to generate a fetch shader will be confined to LGC.

Change-Id: I5c13e256873aaf7ce405a3ef4676a4850eb1dec7