[CI] Add workflow to checking PR formatting with clang-format
This workflow runs clang-format-diff on new pull requests
and checks if the changed lines are formatted. It ignores the
other, existing, code that a PR doesn't modify.

I decided not to use third-party actions, because virtually
all of them use JavaScript and node.js and I wouldn't know how to debug
or customize them. A bash implementation seemed much simpler and
more maintainable to me.