Created new PalMetadata class
The new class (in PalMetadata.cpp and .h) owns the PAL metadata MsgPack
document. This commit does not change anything else -- all the PAL
metadata is still constructed in ConfigBuilderBase and subclasses in the
PatchPreparePipelineAbi pass -- but it allows for future commits to move
towards a model where a PAL metadata entry can be set (via a method in
PalMetadata) at the point it is calculated, instead of having to store
it in an intermediate structure such as IntfData.

Change-Id: I2b9dd4d9a75c3687ead71a53276f8ea8da8a6cdc