Simplified shadow descriptor table option handling
The command-line options for shadow descriptor table handling are now in
the front-end in PipelineContext. LGC now only has to deal with a
pipeline option. That too has been simplified; there is now just a
single pipeline option with a special value to indicate that shadow
descriptor tables are disabled.

In a future commit, we should remove the command-line options entirely,
with an LLPC version bump, and use the pipeline option instead in the
driver.

Change-Id: I116f3fa2dc59ebfa23528be75c670eaac91ce5c2