PatchEntryPointMutate arg allocation overhaul
I did this to isolate the part of the code that scans the user data
nodes so they can be allocated to shader args, and thus SGPRs, so that a
future commit can scan used descriptor sets instead, for when shader
compilation does not have user data layout.

No functional change intended.

Change-Id: I8b1ab0246f212210592215174317d50bdcf3ff68