Use a single pass manager to run all the "whole pipeline" passes.
Previously it was artificially split into two: one pass manager to run
the patching and middle-end optimization passes, and a different one to
run the codegen passes.

A side effect of this change is that the middle-end optimization passes
are run in a pass manager that has access to better target-specific
alias analyses, so some passes like Early CSE are more effective.