Fixed invalid IR from SystemValues::loadDescFromDriverTable
Before this fix, there was a danger that
SystemValues::loadDescFromDriverTable could generate illegal IR (def
after use).

I have not seen the problem in the wild, but only by luck that
SystemValues had already been used to extend a 32-bit pointer to 64-bit
using PC by the time it wants to load a descriptor from the driver
table. I found the problem with a later change that moves descriptor
loading code out of SystemValues, so the problem became more likely.

Change-Id: Ifbefe04aa4a13e8a367305fd0ee5bfb7885df97f