Add lookup table to store layout offsets of primitive shader const buffer
Primitive shader table is implemented as a const buffer (CB). The entries
of this table storing run-time info used by NGG culling. Previously,
layout offsets are calculated when the entries are used. In the future, we
calculate all offsets at the creation time of NGG primitive shader class.

This severs the purpose of allowing PAL to change the CB layout in the
future flexibly. And LGC's actions will be just to change the calculation
in the new function buildPrimShaderCbLayoutLookupTable().

Change-Id: I8f20f895578157d908c52afc8513c8017fa95a4f