Set the afn (ApproxFunc) fast-math flag for float operations.
This is documented as allowing approximate calculations for intrinsics
like sin, log, sqrt etc. In practice it will allow generic middle-end
LLVM IR passes to transform pow(x,y) to exp(y*log(x)) and constant-fold
log(x) when x is a constant, and this doesn't decrease accuracy at all
because AMDGPU backend codegen implements pow with log and exp anyway.

It also allows the backend instruction selector to implement division as
multiplication by the reciprocal. This causes problems for some Vulkan
CTS tests which expect accurate double-precision results, so only set
this flag for half and float, not double.