New vertex fetch pass
This moves vertex fetches out of PatchInOutImportExport and in to a new
VertexFetch pass that runs just before PatchEntryPointMutate. This is a
precursor to implementing a fetch shader in shader compilation and
linking.

The VertexFetch pass is in fact the existing VertexFetch class, which
already implemented all the details of a vertex fetch, but converted
into a pass.

Change-Id: Icf5274d9cfbe3eb59fd15cd150e536157b46a5d2