Set SkipUniformRegions for StructurizeCGF pass
The default for the SkipUniformRegions option for the LLVM StructurizeCFG pass
was recently changed from true to false - this was to avoid bad effects seen
with some compute workloads. However, this change impacted the performance of
some shaders.

This change uses the LLVM option -structurizecfg-skip-uniform-regions to
set the value of SkipUniformRegions to true for shaders, thereby restoring
the performance for the impacted shaders.