Fixed converting immutable sampler detection
ImageBuilder::CreateImageSample has code to detect that its sampler is a
converting immutable sampler, so it can call CreateImageSampleConvert.
However that detection code was completely broken. Fixed.

Also enhanced a comment to point out that a DescriptorYCbCrSampler is a
combined image+sampler where the sampler is ignored and overridden by an
immutable converting sampler. Fixed stride to allow for that.

Change-Id: I9be56eec0feec87624a363505f0cd6d91ea7a034