Add disableLoopUnroll to pipeline shader options
Previously loop unrolling could be disabled by use of the forceLoopUnrollCount
pipeline option with a value of 1. However, this is not applied ito loops that
already have explicit loop metadata (e.g. Unroll or DontUnroll).

The new disableLoopUnroll option is applied to loops whether or not there is
existing metadata. This allows explicit hints to be overridden when appropriate.

The forceLoopUnrollCount=1 option can continue to be used in cases where
existing metadata should be preserved.