[CI] Continue running remaining jobs despite failures.
This disables the default fail-fast behavior which kills all
github actions jobs after a single one fails. The motivation is to
better diagnose failures in the presence of `-Werror` and broken
trunk, like in
https://github.com/GPUOpen-Drivers/llpc/commit/ff90ae47205965cd9f8152ed86da6edeb081f41f.

For example, this allows all gcc builds to continue and report final
status even when clang builds fail.
Relevant docs:
https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstrategyfail-fast.