Stop using confusing ShadowDescriptorTable::Disable inside LGC
ShadowDescriptorTable::Disable was the value of the
shadowDescriptorTable pipeline option to disable shadow descriptors.
Confusingly, I was also using that value in other places in LGC where
the high half of an address is provided to mean "get the high half from
PC".

This commit separates that usage into a new HighAddrPc. For the pipeline
option, it also changes ShadowDescriptorTable::Disable to
ShadowDescriptorTableDisable, as there was no particular need for it to
be an enum class, and it saves some casts.

Change-Id: Ib46c410f24c098d9471798f66f5b050118b72bdf