Force read-only data into .text in full pipeline compile
PAL requires constant data to be in the .text section. LLVM did that by
default for PAL OS type, but a forthcoming LLVM change stops it doing
that unless the global's section name is explicitly set to ".text".

This commit explicitly sets the section of a constant global variable to
".text" for full pipeline compilation. Thus this commit is required
before the LLVM change is landed.

For shader compilation, we want global variables to go into the separate
.rodata section.  The linker then puts them into .text.

Change-Id: Ia5fc13b01f0edc15d9eabad22dedc8f25e234deb