Update test to allow for new InstCombine sinking
Upstream llvm has some additional sinking functionality that requires a change
in this test. The change seems to be an improvement in this case, and it is the
only test that shows a difference (surprisingly).

The instructions that are removed from the check are actually sunk to the bottom
of the shader, but it is tricky to have a test that works before and after the
merge, so I've removed these instructions altogether from the check. I think it
is reasonably safe to do. We could add the instructions back once the merge has
happened if any reviewers think it is worthwhile.