Refactor canUseRelocationShaderElf functions
Two small changes.  First I removed the variable with the return value,
and just do early returns.  This makes the code easier to read.

Second I removed the static variable that would count the number of
pipelines compiled with relocatable shader, and replaced it with a
member variable.  This way the graphics and compute pipelines can share
the counter.  Makes debugging easier.