Add pipeline option to use relocatable shaders.
We want to move to a pipeline option to tell LLPC to use relocable
shaders.  This involves

1) Deprecating the current command line option in llpcCompiler.  I do
not want to remove it yet because that is the only way to currently
enable relocatable shader when using LLPC with an existing xgl driver.
Once xgl has been updated to use the pipeline option this option can be
removed.

2) Adding the pipeline options.

3) Adding a command line option to amdllpc to enable relocatable
shaders.

4) Update tests to use amdllpc's command line option instead of the
depracated one.

Note that we could wait on 3 & 4, and have amdllpc reuse the same
option as before.  I have no strong feelings about that.