Refactor FragColorExport to not use the pipeline state
We do not want the FragColorExport class to depend on the pipeline
state.  Eventually, this class will be used when generating the
color export shader, and the glue shader code should will not have
access to the entire pipeline.

This commit will move some code from FragColorExport to the pipeline
itself.  Getting the color export format is dependent on pipeline
information anyway, so I think it fits better there.

Code that updates the pipeline state for later use, it moved into
PatchInOutImportExport before calling FragColorExport::run.