PatchBufferOp: fix and improve GetBufferDescLength lowering
Since d717e0bd62 ("Hanlde the behavior of getting array length for
a null descriptor"), the implementation of OpArrayLength no longer
subtracts the runtime array's offset, relying on the
GetBufferDescLength implementation to do so. However, that
implementation only did the subtraction in the allowNullDescriptor
case.

Furthermore, the implementation of the allowNullDescriptor is
rather expensive: it shouldn't be necessary to check all dwords
of the descriptor. Instead, just clamping the array length to 0
should do the trick (which has the side effect of likely being
more robust in practice when applications use buffers that are
non-null but too small -- this case seems underspecified in
SPIR-V and Vulkan, so clamping should be fine).

Fixes: d717e0bd62 ("Hanlde the behavior of getting array length for a null descriptor")