Add compute library functionality
A compute library can now be passed into LGC, distinguished from a
compute pipeline by there being no shader entry-point (no call to
Pipeline::markShaderEntryPoint). Functions that need to be visible from
outside are marked in the usual LLVM IR way.

In a compute library, or a compute pipeline with calls,
PatchEntryPointMutate behaves differently: it does not attempt to omit
unused shader inputs, and it passes shader inputs on each call and in to
each function.

A compute library might only know a prefix of the user data layout in
the pipeline that it will eventually be used with. To cater for that, we
always pass 16 sgprs, and we put gl_NumWorkGroups at the start, straight
after the global pointer and shader pointer (i.e. into s[2:3]), and we
always reserve s15 for the spill table pointer, even if it is not
needed.

The "gl_NumWorkGroups at the start" change applies to all compute
shaders, but a standalone compute shader only adds it there if it
actually uses it.

Change-Id: I29b0a6e2230cf16e563aa80ef59d199b1f6995ac