Tidy image and sampler handling in SPIR-V reader
Knowledge of how a SPIR-V image or sampler or pointer thereto translates
to an IR type is now a bit more localized, in preparation for subsequent
commits that will make it more complicated (to support YCbCrSampler).

Use of the Builder calls to get the IR type of a descriptor or
descriptor pointer, and index a descriptor pointer, and load a
descriptor, is now also more localized, in preparation for a subsequent
commit where I am going to tidy the Builder API for descriptors.

Part of this change is to ensure that the IR element type can be
available when doing an index or load, so we can avoid
getPointerElementType and be more compatible with a future LLVM change
to opaque pointer types.

Change-Id: I7695528195c7f50b85e5dec467acdf49c110ec63