Add color export info to the pal metadata
We add information needed to generate the color export instruction to
the pal meta data.  This will be used to create the color export shader
when it is implemented, and will be used to set the CB_SHADER_MASK and
SPI_SHADER_COL_FORMAT entries in the pal metadata when finalizing the
pipeline.