Pass array of image descriptors in to Builder::CreateImageSampleConvert
This is part of the support for multi-plane images in YCbCr conversion.

Builder::CreateImageSampleConvert now allows you to pass an array of
image descriptors, instead of a single one. The implementation of that
function still only uses the first descriptor.

The SPIR-V reader now reads three contiguous image descriptors instead
of a single one when there are any converting samplers in the program.
Most image operations only use the first one, and we rely on later
optimizations to remove the unused second and third ones. But it passes
the whole array to Builder::CreateImageSampleConvert, ready for it to
use them as multiple planes in a future commit.

Change-Id: I061e7fcdc2aa7bfb26d66645782decef509243df