Remove descriptor sizes from GpuProperty
Further to Nicolai's comment on #782, this removes the descriptor sizes
from GpuProperty, and makes them constants in IntrinsDefs.h instead. As
Nicolai says, if the sizes change in some future hardware, there is
likely to be more changes needed than just changing the value in
GpuProperty, and using constants allows better optimization when
building the compiler.

Change-Id: I51c217ae04986853ccdfdcdffd5014b58c4ced5b