Add a shaderdb test for OpImageFetch with dynamic offsets
This patch adds a test for OpImageFetch with dynamic offsets. The input
SPIR-V is based on HLSL compiled with DirectXShaderCompiler
(https://github.com/microsoft/DirectXShaderCompiler).

The test verifies that a loop with N texture fetches with dynamic offsets
results in N image_load instructions (ie. the dynamic texture offset is not
ignored and optimized into a single image load instruction).