Use unnamed fields in builtInUsage and add missing MVG stage mask
It is convenient for us to add new reserved fields if we use unnamed
ones and we don't have to give them "meaningful" names any more.
This change also reserves some fields for future uses.

Also, add missing MVG stage mask to represent a built-in usage that
is valid in vertex shader or geometry shader.

Change-Id: I3203fd1cbb980d6cb4777a14a3e6fdc00681e747