Set minimum waves per eu to 1 instead of 0.
The backend ignores the amdgpu-waves-per-eu attribute if the minimum
value is less than 1 or if the maximum value is greater than the limit
imposed by the hardware architecture. We set the minimum to 0, which
meant that the attribute was always ignored. Setting it to 1 means that
the attribute is not ignored, so we can use the -waves-per-eu command
line option to set a useful maximum.