Fixes for glsl change for != being unordered not ordered fcmp
Changes in glsl (see https://github.com/KhronosGroup/glslang/issues/2259 and
https://github.com/KhronosGroup/glslang/pull/2260) mean that fcmp will now use
une (unordered) rather than one (ordered) ne.

This is in line with most high level languages and means that NaN is handled in
a more expected way.

llpc-lit tests need updating for this change. The change allows for both one and
une during the period of transition.