NGG: Correct ES splitting for deferred vertex export
Originally, we split ES to position export and parameter export. This
was also doned in NGG main path with two basic blocks "expVertPos"
and "expVertParam". However, this is not conceptually correct. Before
NGG culling, we only need position data (pos0) and maybe cull
distance values. And after NGG culling, the exports are done both for
all position values and all parameters.

This is change to correct the mistaken splitting by creating two
corresponding ES-partial functions: one is to fetch cull data while
the other is to do deferred vertex export. Accordingly, NGG main path
is simplified as well ("expVertPos" and "expVertParam" are merged to
"expVert").

The function "mutateEsToVariant" is replaced by "splitEs". The function
"runEsOrEsVariant" is replaced by "runEs" and "runEsPartial".

The change is verified by dEQP-VK.draw.*.

Change-Id: Ib13735d6f14c2817f272d8cea056a37cc474654d