Simplify the use of operandIdx around loadInOutMember()
Previously we use 'operandIdx + 1' to by-pass the first zero offset
introduced from GEP instruction, and we also use 'operandIdx + 1' for
the next level of recursive call to loadInOutMember(). This patch tries
to make the meaning of 'operandIdx' clear, it means the index into
the offset array, and the indexed offset value is against the inOutTy.
To achieve this goal, I just remove the first zero offset contained in
the GEP instruction.