Reimplement dynamic indexed inOut in interpolate functions
The new solution works the same as before in behavior, pre-load all
members for dynamic index, and only load the specified member for static
indexed access. The new solution mainly target to add vector type
support.

I find it hard to extend the previous solution to support various
situations like dynamic-indexed vector or static indexed vector within
a dynamic-indexed array. Here a recursive implementation makes it easy
to insert the 'InsertValue' instructions. And the new solution is easy
to read and maintain.

The reason I choose to implement a new function instead of handle it
inside loadInOutMember() is that would require each data-type with an
extra if-branching for such dynamic-case. So I think it is clean to
keep the new logic in a separate function.

Fixes: dEQP-VK.pipeline.multisample_interpolation.*pushc_component_1*