Add option -strict-unroll-hints for loop metadata
Added new option -strict-unroll-hints with a default value of true.
This default treats loop metadata llvm.loop.unroll.disable and
llvm.loop.unroll.full in the same way as previously - i.e. the hints
will be obeyed as far as is possible.

When -strict-unroll-hints is used with a value of false, then
- llvm.loop.unroll.full is changed to llvm.loop.unroll.enable to
  allow the loop heuristics to determine whether or not to fully unroll
  the loop.
- llvm.loop.unroll.disable is changed to the new target specific loop
  metadata amdgpu.loop.unroll.threshold with a value of 250. This sets
  the starting value for the unroll threshold to a low value so that
  suitably small loops may be unrolled.