Make coding style more consistent. NFC.
Reformat comments with inline documentation to make them consistent
with the most common practice and update the coding standards doc.
Also replace nullptr comparrison with unary negation.

I never remember what the format of these comments is and get confused
by different files using slightly different style.
After this lands, I plan to submit a new PR that will add a github
actions CI check for making sure the @param and @returns comments
adhere to the standard and the source doesn't continue to diverge,
similarly to how the clang-format we've been using for a while.

This was performed by running the following subsitutions in order:
```
1. (\(|\s|\|\||&&)([a-zA-Z0-9]+)\s==\snullptr ==> $1!$2
2. (///?\s+)\\((param)|(returns?)) ==> $1@$2
3. @return\s ==> @returns
4. (///?\s+@param)\s+([a-zA-Z0-9]+)\s+(\[.*\]) ==> $1 $3 $2
5. (///?\s+@param)\s+([a-zA-Z0-9]+)\s+:\s+(\[.*\]) ==> $1 $3 $2 :
6. (///?\s+@param)\s+(\[.*,.*\])\s([a-zA-Z0-9]+) ==> $1 [in/out] $3
7. (///?\s+@param)\s+((\[.*\]\s)?[a-zA-Z0-9]+)\s+([a-zA-Z0-9]+) ==> $1 $3 : \U$4
8. (///?\s+@returns)\s+([a-zA-Z0-9]+) ==> $1 : \U$2
```