Move builtin export generation to the color export shader.
In the initial design, I did not realize that it is possible that the
color export shader could be empty.  This happens when the pipeline
state does not have any color buffer information for the export
instructions that were in the shader.

This causes a problem because we cannot know if the "done" flag should
be set on the builtin export instruction until generating the color
export shader.  This means that the export instruction for the builtin
exports must be generated when building the color export shader.