Don't access the color export state with unlinked shaders
When building a fragment shader that will have a color export shader, we
do not set the color export state in the pipeline state, so we better
not reference it.  Not many changes are needed.  The
"dualSourceBlendEnabled" is was only used when getting the color format
for a location, which is not called when generating a color export
shader.

The "alphaToCoverageEnable" field is only used when generating the pal
metadata for gfx9.  That has been moved to when the pal metadata is
finalized.