Merge all of the RSRC1 values when linking.
When merging the PAL metadata in the elf linker, the RSRC1 field for
each shader is merged by taking the original by taking the bits in the
original value and updating on the VGPR and SGPR entries.  This means
that the remaining entries in the source value are lost.

We fix this by taking the bitwise of the remaining bit.