Merge gpr limits and count correctly.
In the elf linker, we currently merge the gpr count and limit entries in
the PAL metadata by doing a bitwise or operation.  However, that is not
right.  If the fetch shader use `x` sgprs and the main body uses `y`, then
the count should be `max(x,y)`.  The same goes for the limit.