Update the size of symbol related to glue code
When we append or prepend glue code to a shader that changes the size of
the main function.  However, we currently do not update the size of the
symbols.  This add code to the elf linking code that will update the
size of the symbols related to the glue code.