Rework implementation for OpCopyObject
Previously OpCopyObject is handled by a store + load, it will lead to problem when copying pointer type object, for example, mentioned in #834. LLPC fails to handle the pointer-to-pointer type in some lower passes so invalid IR is generated / assertion is triggered.

Directly mapping the llvm value of the Operand to the Result should be enough for OpCopyObject, since the Operand and Result should be exactly the same for OpCopyObject.

This fixes #834.