Fix visitPHINode for PatchBufferOp
When visiting a phi node in PatchBufferOp, an incoming value may be unvisited and result in null when getting it from m_replacementMap. A shader to reproduce this issue is included in the commit.

This commit fixes the issue by recording the values that cannot be gotten from m_replacementMap, then fixup the values after the whole function has been visited.