Link relocatable shaders with data in the .rodata section
The current relocatabale shader linking does not work when there is read
only data.  See https://github.com/GPUOpen-Drivers/llpc/issues/896 for
an example.  So a change is going into LLVM
(https://reviews.llvm.org/D85895) to place the data in the .rodata
section instead of the .text section.  This change is to have the elf
linker handle it correclty.

Fixes #896