Add null image descriptor check for i64 imageLoad
According to GL spec, unused components are filled with (0, 0, 0, 1)
for imageLoad(). Thus, the i64 image returns 1 in W channel. But the
robustness requirement expects a all-zero value when the image
descriptor is null. Thus, we have to add runtime check to handle this
special case since the W channel value is not from HW (hard-coded by
us). HW is unable to handle this.

Change-Id: I23aa29bf1fcd5fca459ada43678e2b994a85d175