Don't use PS_INPUT_* from glue code
The color export shader will have entries for the PS_INPUT_* metadata,
but those values are completely wrong because the inputs to the color
export shader has no connection with the inputs to the pixel shader.  We
need to avoid using those values.

Fixes https://github.com/GPUOpen-Drivers/llpc/issues/861