Add Image resource type workaround
GFX10 hardware cannot handle invalid image resource type [1, 7], so we
change such kind of value to 0 (buffer) to let hardware ignore the request.
This may happen when buggy application declares an image in shader but
provide a buffer descriptor in driver. Also add a pipeline option to disable
it for possible performance loss.