Fix bug causing SPIRV files to be missing for a dumped pipeline.
Address a bug where the SPIRV files would be missing for the first
pipeline dumped if the pipeline dump directory needed to be
created.  This occurs because the SPIRV output was attempted
before the BeginPipelineDump call which creates the dump directory.
Move the SPIRV file output to occur when pipeline data is written
to the .pipe file.