[CI] Remove glslang pin for sanitizer build
The pinned glslang commit in spvgen was updated to a more recent commit.
Building the commit that was pinned in docker lead to failing builds.

I did not do a docker build, but I checked locally that using the old
glslang version leads to the same build failure as in the GitHub CI
and not pinning it works fine.