Fix Google ANGLE issue: vertex fetch failure with unaligned vertex buffer binding offset
For some formats, there is a HW defect when the binding offset of vertex buffer is unaligned, we need to split the vertex fetch operation in this case. But we may not have the vertex buffer binding info at pipeline creation time, we have to split the fetch unconditionally unless there is a better solution.