Don't try to merge ELF binaries containing relocs.
Partial pipeline compilation relies on compiling the shaders in a
pipeline separately and then using llpcElfWriter to merge the ELFs.
Don't do this if the shaders contain global constant data because LLVM
has recently started putting this data in a separate .rodata section and
using relocs in the .text section to refer to it, but llpcElfWriter
doesn't know how to update the relocs when it merges the ELFs.