Fix-ups for now deprecated getNumElements on VectorType
Changes in LLVM mean that VectorType can have a variable number of elements that
are only defined at run-time.

getNumElements is now deprecated for VectorType and instead, the value should be
cast to FixedVectorType and then getNumElements can be used meaningfully. All
vector types used by LLPC are in reality FixedVectorType, so the cast can be
done safely.