Fuzzer test no longer fails with pending upstream commit
The test is looking for failures when branch offset is too large for immediate
field.
Upstream change has made the shader shorter so it now always passes.

Removing the test as it doesn't do what it is supposed to any more. We'll need a
new fuzzer test if we want to replicate this.