Remove GraphicsFuzz_ReplaceAllUsesWithNotValid.spvasm no longer works
Removing this test as a more recent LLVM fails to exhibit the issue. I think
that the issue is probably still there, but hidden by changes upstream.

Since there's already an issue filed for this (with a reproducer for the noted
versions of the source), I'm going to remove the test to allow merges and
promotions.

My assumption is that a test will be reinstated (that passes) once the original
issue is resolved.