Refactor shadow desc table code
We want to use relocation so that unlinked shaderd do not have to access
those options.  This change will refactor a few thing to make that
easier.

- Minor refactoring so that the effects of accessing the shadow desc
option are short lived.  We don't want to set a variable that has a long
live range.  This removes the `shadow` parameter from
`DescBuilder::getDescPtr`.

- `AddressExtender::entend` is modified to accept the high half of the
addess as a value instead of an integer.  This will be useful when the
high half is obtained from a relocation.