Fix hashing of entire pipeline.
The location that is changed is suppose to compute the hash for the
entire pipeline, and that is fed back in the pipeline elf.  We should
not be skipping parts of the pipeline because they are not used for the
relocatable shader compilation.

No test was added because I do not know how to write a stable test
checking that the hash is correct.