NGG: Add runtime passthrough support
Sometimes, we couldn't determine if NGG culling mode should be disabled
at compilation time. For example, if total vertex count in a subgroup
is below 16, such small subgroup shouldn't perform culling. In the
future, certain graphics states are dynamically set. We rely on part
of them to determine the feasibility of NGG culling. Therefore, a
runtime passthrough path is imperative.

This change adds such support. We define small subgroup as a compilation
option exposed to tuning: -ngg-small-subgroup-threshold.

Change-Id: Ife28281e2587c441c8d930d51d1ba568435ee307