Fix for effects of upstream LLVM change
An upstream LLVM change increased the default block size threshold for inner
loop analysis which resulted in a big increase in compile time, and additional
spilling for a small number of pathological cases.

As the upstream change (https://reviews.llvm.org/D86248) does not benefit any
graphics shaders, the fix is for LGC to simply override the value back to the
previous default value of 20.