Amend loop metadata handling
Change the use of amdgpu.loop.unroll.threshold to llvm.loop.unroll.threshold
to match the change in the metadata name following the LLVM code review.

Also add an extra command line option so that the handling of the DontUnroll
and Unroll loop hints can be controlled independently to provide more
flexibility.