Generate readfirstlane efficiently for uniform case
The Vulkan Enviroment for SPIR-V (section "Validation Rules within a
Module",  version 1.2.140) says that:
"If an instruction loads from or stores to a resource (including atomics
and image instructions)
and the resource descriptor being accessed is not dynamically uniform,
then the operand
corresponding to that resource (e.g. the pointer or sampled image
operand) must be decorated
with NonUniform."
The usage of NonUniform decoration is modified according to spec both in
CTS and glslang.
Example:
layout(set = 0, binding = 0) uniform texture2D uTex[];
layout(set = 1, binding = 0) uniform sampler uSamp;
layout(location = 0) flat in int Index;
layout(location = 0) out vec4 FragColor;
void main()
{
  // Old usage not according spec
  FragColor =
  texture(nonuniformEXT(sampler2D(uTex[Index],uSamp)),vec2(0.5));
  // New usage according spec.
  FragColor =
  texture(nonuniformEXT(sampler2D(uTex[Index],uSamp)),vec2(0.5));
}
It produces the following SPIR-V (only extract provided):
// Old usage
...
OpDecorate %18 NonUniform
OpDecorate %21 NonUniform
...
%5 = OpLabel
%17 = OpLoad %int %Index
%18 = OpCopyObject %int %17 <-- The index is NonUniform
%20 = OpAccessChain %_ptr_UniformConstant_10 %uTex %18
%21 = OpLoad %10 %20             <-- The image itself is
NonUniform
%25 = OpLoad %22 %uSamp
%27 = OpSampledImage %26 %21 %25
%31 = OpImageSampleImplicitLod %v4float %27 %30 <-- The
sampled image is *not* NonUniform
...
// New usage
...
OpDecorate %33 NonUniform
...
%10 = OpTypeImage %6 2D 0 0 0 1 Unknown
...
%31 = OpTypeSampledImage %10
...
%32 = OpSampledImage %31 %20 %30
%33 = OpCopyObject %31 %32
%37 = OpImageSampleImplicitLod %7 %33 %36
...
Fixes:
-https://github.com/GPUOpen-Drivers/llpc/issues/840
-dEQP-VK.descriptor_indexing.sampled_image*.
The reason of cts failutres is that the descriptor index
is mistook as uniform in indexDescPtr which it should be
non-uniform.
Solution:
-Remove the non-uniform flag in indexDescPtr, which can make the test
pass.
-Generate readfirstlane for operands without non-uniform decoration.
1)Insert readerfirstlane on the resource descriptor
initially
2)Add a otpimization pass in middle-end to push up
readfirstlane to transform the use of vgpr to sgpr  as
many as possible.